requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fast-bss-eval==0.1.4
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
gitpython==3.1.40
google-auth-oauthlib==1.0.0
google-auth==2.23.2
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
huggingface-hub==0.17.3
hyperpyyaml==1.2.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
jupyter-client==8.3.1
jupyter-core==5.3.1
keras==2.14.0
kiwisolver==1.4.5
lazy-loader==0.3
libclang==16.0.6
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.41.1
markdown-it-py==3.0.0
markdown==3.5
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.7.3
mdurl==0.1.2
mido==1.3.0
ml-dtypes==0.2.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
music21==8.3.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.58.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.0.0
pip==23.2.1
platformdirs==3.10.0
pooch==1.8.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycocotools==2.0.7
pycparser==2.21
pydub==0.25.1
pygments==2.16.1
pyparsing==3.1.1
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==2.1.1
pytorch-ranger==0.1.1
pytz==2023.3.post1
pywin32==304
pyyaml==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests-oauthlib==1.3.1
requests==2.31.0
rich==13.6.0
rsa==4.9
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.5
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.15
stack-data==0.6.2
sympy==1.12
tensorboard-data-server==0.7.1
tensorboard==2.14.1
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.14.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.8
tokenizers==0.14.1
torch-complex==0.4.3
torch-mir-eval==0.4
torch-optimizer==0.3.0
torch==2.2.0.dev20230913
torchaudio==2.2.0.dev20230913
torchmetrics==1.2.0
torchvision==0.17.0.dev20230913
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.34.1
typing-extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.202
urllib3==2.0.4
wandb==0.16.0
wcwidth==0.2.6
webcolors==1.13
werkzeug==3.0.0
wheel==0.41.2
win-inet-pton==1.1.0
wrapt==1.14.1
yarl==1.9.2
zipp==3.16.2